Rename new journal entry form for better readability
